Introduce min/max variance, adapt initSurface, modify configuration change handling
A new option to vary the min and max wind speed of individual surfaces is introduced. This means that new surfaces won't necessarily have the exact same min and max wind speed. With mods that provide new surfaces this might lead to some variation. To properly handle these changes the initSurface block was moved and modified to respect the new settings and is being called in onConfigurationChanged. In general after changing settings the surfaces will be reintialized. New values will be used when a new season begins.